requirements.txt
# MangaAutoColor Pro - Dependências

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.27.0
transformers>=4.35.0
accelerate>=0.24.0
safetensors>=0.4.0
peft>=0.7.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.22.0
scipy>=1.11.0

# Detection
ultralytics>=8.0.200

# Segmentation (ADR 004)
sam2>=1.1.0

# Vector Database
faiss-cpu>=1.7.4; platform_system!="Windows" or python_version>="3.10"
faiss-gpu>=1.7.4; platform_system=="Linux"

# Data Processing
# NumPy < 2.0 obrigatório para compatibilidade com PyTorch 2.4.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
pyarrow>=14.0.0
scikit-learn>=1.3.0

# Face Recognition / Identity
insightface>=0.7.3
# ONNX Runtime GPU for CUDA acceleration (Linux and Windows)
onnxruntime-gpu>=1.16.0

# Image Processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
tqdm>=4.66.0

# Web Interface
gradio>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Configuration
pyyaml>=6.0.1
pydantic>=2.5.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
huggingface-hub>=0.19.0
einops>=0.7.0

# Tokenizers / Serialization
protobuf>=3.20.0
sentencepiece>=0.1.99

# Visualization (debug)
matplotlib>=3.7.0

# Optional optimizations
xformers>=0.0.22; platform_system=="Linux"

# Development
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0
black>=23.0.0
flake8>=6.0.0
# FastAPI dependencies
python-multipart
